Refactor 1; split files into modules.
